qcacld-3.0: Parse device info nl attribute in the oem vendor cmd
Userspace uses enum qca_vendor_oem_device_type to indicate
device type as vdev or pdev in the oem data vendor nl command.

This attribute is set to 1 to indicate that the oem data command
is meant for pdev and it is 0 for vdev. The default is vdev when
this nl attribute is not present.

The host driver should update the pdev_id of the vdev and fill
the pdev_vdev_flag to 1 for pdev commands.

Change-Id: I8bdf436759e0935c52bb09c2865fcb4a7a5ee4fc
CRs-Fixed: 2693892
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>