qcacld-3.0: Overide the value of sta sap SCC DFS ini
Currently the driver allows standalone SAP and
STA+SAP SCC even if the dfs master mode is 0
and the ini g_sta_sap_scc_on_dfs_chan if
configured 2 assumes that the dfs master mode is
enabled.
If the user configures master mode as 0 then the
ini g_sta_sap_scc_on_dfs_chan should not be 2
since standlaone SAP can come up on DFS channel
according to the implementation.

Fix is to overide the value of g_sta_sap_scc_on_dfs_chan
to zero if the master mode is disabled.

Change-Id: Ife02a3545a5df7dfa7855feea7965628a1f513a3
CRs-Fixed: 2713392
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>