qcacld-3.0: Change legacy macro msg types to enum msg types
Driver maintains a global list of messages (qdf trace) which
different modules posts to each other. If required, this list
can be processed with the help of post processing cmm script.

Currently some of the message types are defined as macros and
these macros can not be used directly in the post processing
cmm script. Enums are much more easy to process with the help
of the cmm script.

To address above issue below fixes are done with this change:
1. Assign enum name to pe message types, lim message types,
   sme message type.
2. Change MLM message type from macros to enum type.
3. Change WMA message type from macros to enum type.

Change-Id: Ibd91a5371457a542cd141f680dda44a48b6f7d98
CRs-Fixed: 2719698
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>