qcacld-3.0: Defer recovery cleanup to shutdown
As part of the firmware down notification if the indication comes
in the non-interrupt context the driver flushes the idle timeout
work and purges the shutdown down notifier.

If the the idletimeout is already running, as part of the idletimeout
work the platform,mhi irq's will also be freed, if the firmware crashes
and enters the RDDM mode during this time, platform driver will notify
the host through the uevent. Since free_irq waits for all the pending
interrupts to be serviced resulting in the deadlock.

Dont flush the work for the pci bus so the interrupt could
be completed and free irq could also be unblocked.

Change-Id: I4e106ff774fa90544ba97737a87682f42f9288df
CRs-Fixed: 2720348
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>