qcacld-3.0: Use channel freq apis instead of channel num apis
Currently for get channel list command host driver is using
channel number based APIs to get the channels state which may
result in incorrect state for some of the channels of 4.9GHz
channels as 4.9GHz channels are overlapping with the 2.4 GHz
channels.

To address above issue use the frequency based apis to get the
channel state.

Change-Id: I3d36c5aa73e42f6a6e4d82e1b7f35412b0c600c2
CRs-Fixed: 2721078
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>