qcacmn: Modularize finding radar affected sub channel function
The radar affected sub channel are determined by the bangradar type,
whether sub-channel marking is enabled or not, the center frequency
of the band on which the radar was detected. This block of code is present
in processing radar indication function and can be modularized.

Move the logic to find radar affected sub-channels into a single function.

Change-Id: Ic0a37fdfa97cbc1bcd2b9a23fd6642f22b43a733
CRs-Fixed: 2737120
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>